requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels @ git+https://github.com/django/channels.git@643d0832698d5306c01927add7b4aa34da1c457d
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
distlib==0.3.8
Django==5.1.1
djangorestframework==3.15.2
filelock==3.15.4
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
kombu==5.4.0
msgpack==1.0.8
oauthlib==3.2.2
pbr==6.1.0
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.1.0
setuptools==74.1.2
six==1.16.0
sqlparse==0.5.1
stevedore==5.3.0
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
virtualenv==20.26.3
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
wcwidth==0.2.13
websocket-client==1.8.0
zope.interface==7.0.3
